# src/main/resources/api/api.yaml
openapi: 3.0.3
info:
  title: Sample API
  version: v1.0
servers:
  - url: /
paths:
  /books:
    get:
      summary: Get all books
      description: Get all books
      operationId: getAllBooks
      parameters:
        - in: query
          name: onlyLeapYears
          required: false
          schema:
            type: boolean
            default: false
          description:
            If true, only return books published in leap years.  If false, only
            return all books
      responses:
        "200":
          description: The query results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./schemas/book.yaml"
    post:
      summary: Create book
      description: Create a new book
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/bookForCreation.yaml"
      responses:
        "201":
          description: The created book
          content:
            application/json:
              schema:
                $ref: "./schemas/book.yaml"
  /books/{bookId}:
    get:
      summary: Get a book
      description: Get a single book
      operationId: getBook
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The found book results
          content:
            application/json:
              schema:
                $ref: "./schemas/book.yaml"
        "404":
          description: The book was not found
    put:
      summary: Update a book
      description: update a book using a specific book ID
      operationId: updateBook
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/bookForCreation.yaml"
      responses:
        "200":
          description: Book updated!
          content:
            application/json:
              schema:
                $ref: "./schemas/book.yaml"
        "404":
          description: The book was not found